+++ js/util.js
@@ -1,0 +1,22 @@
+//функция получения целого числа
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+//возвращает случайный элемент из массива
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+const createIdGenerator = () => {
+  let lastGeneratedID = 0;
+
+  return () => {
+    lastGeneratedID += 1;
+    return lastGeneratedID;
+  };
+};
+
+export { getRandomInteger, getRandomArrayElement, createIdGenerator };
